Added js file for the DOM. #1
